create file for adjust shape function, remove recently created parameters
