+++ neurolib/models/fhn/timeIntegration.py
@@ -2,6 +2,7 @@ import numpy as np
 import numba
 
 from . import loadDefaultParams as dp
+from ...utils import adjust_params as ap
 
 
 def timeIntegration(params):
@@ -82,16 +83,13 @@ def timeIntegration(params):
     x_ou = params["x_ou"]
     y_ou = params["y_ou"]
 
-    x_ext_const = params["x_ext_const"]
-    y_ext_const = params["y_ext_const"]
-
     # state variable arrays, have length of t + startind
     # they store initial conditions AND simulated data
     xs = np.zeros((N, startind + len(t)))
     ys = np.zeros((N, startind + len(t)))
 
-    x_ext = adjust_shape(params["x_ext"], xs)
-    y_ext = adjust_shape(params["y_ext"], ys)
+    x_ext = ap.adjust_shape(params["x_ext"], xs)
+    y_ext = ap.adjust_shape(params["y_ext"], ys)
 
     # ------------------------------------------------------------------------
     # Set initial values
@@ -139,8 +137,6 @@ def timeIntegration(params):
         ys,
         xs_input_d,
         ys_input_d,
-        x_ext_const,
-        y_ext_const,
         x_ext,
         y_ext,
         alpha,
@@ -178,8 +174,6 @@ def timeIntegration_njit_elementwise(
     ys,
     xs_input_d,
     ys_input_d,
-    x_ext_const,
-    y_ext_const,
     x_ext,
     y_ext,
     alpha,
@@ -235,15 +229,13 @@ def timeIntegration_njit_elementwise(
                 - ys[no, i - 1]
                 + xs_input_d[no]  # input from other nodes
                 + x_ou[no]  # ou noise
-                + x_ext_const[no]  # constant external input
-                + x_ext[no, i] # time-dependent external input
+                + x_ext[no, i] # external input
             )
             y_rhs = (
                 (xs[no, i - 1] - delta - epsilon * ys[no, i - 1]) / tau
                 + ys_input_d[no]  # input from other nodes
                 + y_ou[no]  # ou noise
-                + y_ext_const[no]  # constant external input
-                + y_ext[no, i] # time-dependent external input
+                + y_ext[no, i] # external input
             )
 
             # Euler integration
@@ -255,46 +247,3 @@ def timeIntegration_njit_elementwise(
             y_ou[no] = y_ou[no] + (y_ou_mean - y_ou[no]) * dt / tau_ou + sigma_ou * sqrt_dt * noise_ys[no]  # mV/ms
 
     return t, xs, ys, x_ou, y_ou
-
-
-def adjust_shape(original, target):
-    """
-    Tiles and then cuts an array (or list or float) such that
-    it has the same shape as target at the end.
-    This is used to make sure that any input parameter like external current has
-    the same shape as the rate array.
-    """
-
-    # make an ext_exc_current ARRAY from a LIST or INT
-    if not hasattr(original, "__len__"):
-        original = [original]
-    original = np.array(original)
-
-    # repeat original in y until larger (or same size) as target
-
-    # tile until N
-
-    # either (x,) shape or (y,x) shape
-    if len(original.shape) == 1:
-        # if original.shape[0] > 1:
-        rep_y = target.shape[0]
-    elif target.shape[0] > original.shape[0]:
-        rep_y = int(target.shape[0] / original.shape[0]) + 1
-    else:
-        rep_y = 1
-
-    # tile once so the array has shape (N,1)
-    original = np.tile(original, (rep_y, 1))
-
-    # tile until t
-
-    if target.shape[1] > original.shape[1]:
-        rep_x = int(target.shape[1] / original.shape[1]) + 1
-    else:
-        rep_x = 1
-    original = np.tile(original, (1, rep_x))
-
-    # cut from end because the beginning can be initial condition
-    original = original[: target.shape[0], -target.shape[1] :]
-
-    return original
